demo/requirements.txt
Pillow
Flask
tensorflow>=2.0.0
pandas
numpy